Move drop logic into kernel
